Change unique identification functions to use cell.Checksum()
Elements that implement the `cell` interface now have to implement
a Checksum method which give us a pseudo unique representation of the
data.

In the implementation of the default types I've opted for seeding the
stringer representation of the value with the type. i.e.: A String type
containing a string `s = "1"` will generate a checksum like so:

    md5(s+"String")

whereas for an Int whose string representation is `s = "1"` will be:

    md5(s+"Int")

This causes that when we compare different columns in a DataFrame we
will not treat an Int and String column as equal even if they have the
same string representation.