Add new column type: Bool
Bool represents boolean elements, it also allows for a nil pointer
representing the NA value, as parsing errors could lead to it.

A consequence of this change is that now all cells should implement
the ToBool() interface so that we are sure we can cast cells to other
types properly.