Add `column` methods to identify contained NA elements
Method col.hasNa() will simply identify if any of the elements inside
the column is NA, whereas col.na() will run generate a []bool where
each of it's elements will be true or false depending on it being NA
or not.

In addition to this, a bug has been fixed where `Bool` type constructor
`Bools()` was not parsing `bool` and `[]bool` elements correctly.