Add Names LoadOption to set initial column names
Now can now give initial column names while you load/read the data. I choose the use the same behaviour as you get with `pandas`. So when you specify too many column names you’ll receive an error. And when you specify not enough column names, the givens names will be applied to the left most columns. Any remaining unnamed columns will be given a value by the `fixColNames` function.

Replaced all uses of `%v` with the appropiate formatter where applicable, to be as exact as possible.

And ordered the load options so they (the functions to set the options and the structs fields where they are declared and used) are all in the same order. The ordering change is purely cosmetic to satisfy my OCD :wink: